Switch to using HAS_XAOD flag instead of ASG_RELEASE to look for if xAOD EDM is available in the setup when compiling custom observables
It was realized that ASG_RELEASE isn't longer available in CMake based releases of AnalysisBase.
The custom observables were returning with 0 as a consequence - super annoying to find this was the cause of a couple hours of confusion!
We should revisit this issue in the near future and figure out how to provide the proper treatment.